Split up file
Split the file into 3 files
main window
rotation panel and viewer controller
